+++ apps/autotest/scripts/rootModuleTest.js
@@ -1,3 +1,3 @@
 module.exports.root = function () {
-  console.log('require("/scripts/rootModuleTest") works, so NODE_PATh is set to configPath')
+  console.log('require("/scripts/rootModuleTest") works, so NODE_PATH is set to configPath')
 }